Extracted turret look/rotate logic into a helper class (W.I.P).
