+++ src/main/java/org/avp/block/BlockTurret.java
@@ -10,7 +10,6 @@ import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.item.EntityItem;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
@@ -20,7 +19,6 @@ import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.Explosion;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
@@ -62,19 +60,6 @@ public class BlockTurret extends Block
     public boolean hasTileEntity(IBlockState state)
     {
         return true;
-    }
-
-    @Override
-    public void onBlockPlacedBy(World world, BlockPos pos, IBlockState state, EntityLivingBase placer, ItemStack stack)
-    {
-        super.onBlockPlacedBy(world, pos, state, placer, stack);
-
-        TileEntityTurret tile = (TileEntityTurret) world.getTileEntity(pos);
-
-        if (tile != null)
-        {
-            tile.setDirection(MathHelper.floor(((placer.rotationYaw * 4F) / 360F) + 0.5D) & 3);
-        }
     }
 
     @Override
