+++ src/main/java/org/avp/packets/client/PacketTurretSync.java
@@ -30,7 +30,7 @@ public class PacketTurretSync implements IMessage, IMessageHandler<PacketTurretS
         this.y = turret.getPos().getY();
         this.z = turret.getPos().getZ();
         this.targets = turret.getTargetListTag();
-        this.rotation = turret.getRotation();
+        this.rotation = turret.getLookHelper().getRotation();
     }
 
     @Override
